// src/app/models/chromeModel.ts
import { SafeHtml } from '@angular/platform-browser';

export class Chrome {
    id: number;
    title: string;
    album: string;
    img: string [];
    price: string;
    html: string;
    htmlSafe: SafeHtml;
    up:string;
    dw: string;
    show:boolean = false;
    shopping: boolean = false;
}